refactor: rename ltrace-decoder to tacit-decoder and update related dependencies

+++ src/frontend/br_mode.rs
@@ -6,8 +6,8 @@ pub enum BrMode {
     BrReserved = 0b11,
 }
 
-impl From<u8> for BrMode {
-    fn from(value: u8) -> Self {
+impl From<u64> for BrMode {
+    fn from(value: u64) -> Self {
         match value {
             0b00 => BrMode::BrTarget,
             0b01 => BrMode::BrHistory,
